requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-storages==1.14.3
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
kubernetes==30.1.0
oauthlib==3.2.2
packaging==24.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
